% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_googlesheet_request.R
\name{add_fake_sheet}
\alias{add_fake_sheet}
\title{add_fake_sheet}
\usage{
add_fake_sheet(fake_sheet, sheet_name)
}
\arguments{
\item{fake_sheet}{fake sheet dribble created with fake_gs4_dribble}

\item{sheet_name}{name of the new sheet}
}
\description{
Adds a fake sheet to a fake_sheet dribble.
}
\examples{
library(tablespan)
fake_sheet <- fake_gs4_dribble()
add_fake_sheet(fake_sheet, sheet_name = "new_sheet")
}
