% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_gt.R
\name{add_gt_footnote}
\alias{add_gt_footnote}
\title{add_gt_footnote}
\usage{
add_gt_footnote(gt_tbl, footnote)
}
\arguments{
\item{gt_tbl}{gt table}

\item{footnote}{footnote text}
}
\value{
gt
}
\description{
Add a footnote to a gt table
}
\keyword{internal}
