% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_gt.R
\name{add_gt_spanner_partial}
\alias{add_gt_spanner_partial}
\title{add_gt_spanner_partial}
\usage{
add_gt_spanner_partial(gt_tbl, tbl_partial)
}
\arguments{
\item{gt_tbl}{gt table without spanners}

\item{tbl_partial}{left or right hand side header of a table created with
tablespan::tablespan}
}
\description{
Adds the spanners of the left hand side or right hand side of the headers
defined in tablespan table to a gt table.
}
\keyword{internal}
