% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_gt.R
\name{add_gt_spanners}
\alias{add_gt_spanners}
\title{add_gt_spanners}
\usage{
add_gt_spanners(gt_tbl, tbl)
}
\arguments{
\item{gt_tbl}{gt table without spanners}

\item{tbl}{table created with tablespan::tablespan}
}
\description{
Adds the spanners defined in a tablespan table to a gt table.
}
\keyword{internal}
