% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_gt.R
\name{add_gt_titles}
\alias{add_gt_titles}
\title{add_gt_titles}
\usage{
add_gt_titles(gt_tbl, title, subtitle)
}
\arguments{
\item{gt_tbl}{gt table}

\item{title}{title text}

\item{subtitle}{subtitle text}
}
\value{
gt
}
\description{
Add a title and subtitle to a gt table
}
\keyword{internal}
