% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicTables.R
\name{add_header_level}
\alias{add_header_level}
\title{add_header_level}
\usage{
add_header_level(parsed_partial)
}
\arguments{
\item{parsed_partial}{The left hand side or right hand side of the parsed
table}
}
\value{
the parsed_partial with additional level fields
}
\description{
basicTables represents headers as (highly) nested lists. To determine the level
at which each entry resides, we have to get the number of root elements below each
parent element. For example, in the following table, x is on level 2, but x1,
x2, y1, and y2 are at level 1:
}
\details{
\preformatted{
|    x    |
| x1 | x2 | y1 | y2}

add_header_level adds the level to each element in a table header.
}
\examples{
library(basicTables)
deparsed <- basicTables:::deparse_formula(formula =
 (`Row Name` = `Row 1` + `Row 2`) ~ `Column 1` + (`Column Banner` = `Column 2` + `Column 3`))
str(deparsed)

deparsed <- basicTables:::add_header_level(deparsed$rhs)
str(deparsed)
deparsed$level
}
\keyword{internal}
