% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicTables.R
\name{add_header_width}
\alias{add_header_width}
\title{add_header_width}
\usage{
add_header_width(parsed_partial)
}
\arguments{
\item{parsed_partial}{The left hand side or right hand side of the parsed
table}
}
\value{
the parsed_partial with additional width fields
}
\description{
basicTables represents headers as (highly) nested lists. To determine how
wide each entry in the header must be (i.e., how many cells it will get in
the Excel output), we have to get the number of root elements each parent element
spans. For example, in the following table, x spans two elements x1 and x2:
\preformatted{
|    x    |
| x1 | x2 |}
}
\details{
add_header_width adds the span to each element in a table header.
}
\examples{
library(basicTables)
deparsed <- basicTables:::deparse_formula(formula =
 (`Row Name` = `Row 1` + `Row 2`) ~ `Column 1` + (`Column Banner` = `Column 2` + `Column 3`))
str(deparsed)

deparsed <- basicTables:::add_header_width(deparsed$rhs)
str(deparsed)
deparsed$width
}
\keyword{internal}
