% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_googlesheet.R
\name{as_googlesheet}
\alias{as_googlesheet}
\title{as_googlesheet}
\usage{
as_googlesheet(
  tbl,
  google_sheet,
  sheet = "Table",
  start_row = 1,
  start_col = 1,
  merge_rownames = TRUE,
  silent = FALSE
)
}
\arguments{
\item{tbl}{table created with tablespan::tablespan}

\item{google_sheet}{Google Sheet spreadsheet dribble created with googlesheets4::gs4_get()}

\item{sheet}{name of the sheet to which the table should be written to}

\item{start_row}{row at which to start the table}

\item{start_col}{column at which to start the table}

\item{merge_rownames}{should row names with identical entries be merged?}

\item{silent}{if FALSE, messages about the writing process will be shown}
}
\value{
Google Sheet spreadsheet dribble that can be further edited or saved
}
\description{
Write a tablespan table to a Google Sheet.
}
\examples{
\dontrun{
library(tablespan)
library(dplyr)

# First summarize the data:
summarized_table <- mtcars |>
  group_by(cyl, vs) |>
  summarise(N = n(),
            mean_hp = mean(hp),
            sd_hp = sd(hp),
            mean_wt = mean(wt),
            sd_wt = sd(wt))

# Now, create a table with grouping variables as row names and spanners:
tbl <- tablespan(data = summarized_table,
          formula = Cylinder:cyl + Engine:vs ~
            N +
            (`Horse Power` = Mean:mean_hp + SD:sd_hp) +
            (`Weight` = Mean:mean_wt + SD:sd_wt),
          title = "Motor Trend Car Road Tests",
          subtitle = "A table created with tablespan",
          footnote = "Data from the infamous mtcars data set.")

if(require_googlesheets4(throw = FALSE)) {
  # Get the Google Sheet (replace with your actual sheet URL)
  google_sheet <- googlesheets4::gs4_get(ss = "link-to-your-google-sheet")

  # Write the table to the sheet
  as_googlesheet(tbl = tbl, google_sheet = google_sheet)
}
}
}
