% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_hux.R
\name{as_hux}
\alias{as_hux}
\title{as_hux}
\usage{
as_hux(tbl)
}
\arguments{
\item{tbl}{table created with tablespan::tablespan}
}
\value{
huxtable that can be further adapted with the gt package.
}
\description{
Translates a table created with tablespan to a huxtable. See <https://hughjonesd.github.io/huxtable/index.html>.
}
\details{
Huxtable is an extremely versatile table creator for R. Once translated to a huxtable, the tablespan table
is easy to export to all formats directly supported by huxtable.
}
\examples{
library(tablespan)
library(dplyr)
data("mtcars")

summarized_table <- mtcars |>
  group_by(cyl, vs) |>
  summarise(N = n(),
            mean_hp = mean(hp),
            sd_hp = sd(hp),
            mean_wt = mean(wt),
            sd_wt = sd(wt))

tbl <- tablespan(data = summarized_table,
                 formula = (LHS = Cylinder:cyl + Engine:vs) ~
                   N +
                   (Results = (`Horse Power` = Mean:mean_hp + SD:sd_hp) +
                      (`Weight` = Mean:mean_wt + SD:sd_wt)))
if(require_huxtable(throw = FALSE)){
  hux_tbl <- as_hux(tbl)
  hux_tbl
}
}
