% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_string.R
\name{as_string}
\alias{as_string}
\title{as_string}
\usage{
as_string(tbl, digits = 2, n = 3, ...)
}
\arguments{
\item{tbl}{result from tablespan}

\item{digits}{number of digits to round doubles to}

\item{n}{number of rows to print
to print the tablespan table. This allows for styling to be printed}

\item{...}{additional arguments passed to prmatrix or huxtable (if use_hux = TRUE)}
}
\value{
nothing
}
\description{
as_string
}
\examples{
library(tablespan)
library(dplyr)
data("mtcars")

summarized_table <- mtcars |>
  group_by(cyl, vs) |>
  summarise(N = n(),
            mean_hp = mean(hp),
            sd_hp = sd(hp),
            mean_wt = mean(wt),
            sd_wt = sd(wt))

tbl <- tablespan(data = summarized_table,
                 formula = (LHS = Cylinder:cyl + Engine:vs) ~
                   N +
                   (Results = (`Horse Power` = Mean:mean_hp + SD:sd_hp) +
                      (`Weight` = Mean:mean_wt + SD:sd_wt)))
cat(as_string(tbl))
}
