% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicTables.R
\name{bt}
\alias{bt}
\title{bt}
\usage{
bt(data, formula)
}
\arguments{
\item{data}{data set}

\item{formula}{formula to create table}
}
\value{
list
}
\description{
Create a basic table
}
\examples{
data("iris")
bt(data = iris[iris$Species == "setosa", ],
   formula = Species ~ (Sepal = Sepal.Length + Sepal.Width) +
     (Petal = Petal.Length + Petal.Width))
}
