% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bt_styles.R
\name{bt_styles}
\alias{bt_styles}
\title{bt_styles}
\usage{
bt_styles(
  background_style = openxlsx::createStyle(fgFill = "#ffffff"),
  hline_style = openxlsx::createStyle(border = "Bottom", borderColour =
    openxlsx::openxlsx_getOp("borderColour", "black"), borderStyle =
    openxlsx::openxlsx_getOp("borderStyle", "double")),
  vline_style = openxlsx::createStyle(border = "Right", borderColour =
    openxlsx::openxlsx_getOp("borderColour", "black"), borderStyle =
    openxlsx::openxlsx_getOp("borderStyle", "double")),
  title_style = openxlsx::createStyle(fontSize = 14, halign = "left", textDecoration =
    "bold"),
  subtitle_style = openxlsx::createStyle(fontSize = 14, halign = "left", textDecoration =
    "bold"),
  header_style = openxlsx::createStyle(fontSize = 14, halign = "center", border =
    "BottomLeftRight", borderColour = openxlsx_getOp("borderColour", "black"),
    borderStyle = openxlsx_getOp("borderStyle", "double"), textDecoration = "bold"),
  footnote_style = openxlsx::createStyle(fontSize = 12, halign = "left")
)
}
\arguments{
\item{background_style}{color etc. for the entire background of the table}

\item{hline_style}{style for the horizontal lines used in the table. Note:
the style for the lines under spanners is defined in the title_style.}

\item{vline_style}{style for the vertical lines used in the table. Note:
the style for the lines under spanners is defined in the title_style.}

\item{title_style}{style applied to the table subtitle}

\item{header_style}{style applied to the table header (column names)}

\item{footnote_style}{style applied to the table footnote}
}
\description{
Define styles for different elements of the table.
}
\examples{
bt_styles()
}
