% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablespan.R
\name{check_variables}
\alias{check_variables}
\title{check_variables}
\usage{
check_variables(data, variables)
}
\arguments{
\item{data}{data set}

\item{variables}{variable names from formula}
}
\description{
Checks if the variables specified in the table formula exist in the data set
}
\keyword{internal}
