% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabelle.R
\name{construct_header}
\alias{construct_header}
\title{construct_header}
\usage{
construct_header(deparsed)
}
\arguments{
\item{deparsed}{table formula translated in nested list}
}
\value{
deparsed with widht and levels fields
}
\description{
Adds width and depth (number of levels) to the deparsed
table description.
}
\keyword{internal}
