% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_table.R
\name{create_outlines}
\alias{create_outlines}
\title{create_outlines}
\usage{
create_outlines(tbl, workbook, sheet, locations, styles)
}
\arguments{
\item{tbl}{table created with tabelle::tabelle}

\item{workbook}{Excel workbook created with openxlsx::createWorkbook()}

\item{sheet}{name of the sheet to which the table should be written to}

\item{locations}{list with overview of row and col locations for different table elements}

\item{styles}{openxlsx style for the different table elements (see ?tabelle::tab_styles).
The styles element also allows applying custom styles to parts of the data shown in the
table body.}
}
\description{
Adds vertical and horizontal bars to the table.
}
\keyword{internal}
