% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styles.R
\name{create_style_column}
\alias{create_style_column}
\title{create_style_column}
\usage{
create_style_column(
  format,
  background_color = "#ffffff",
  text_color = "#000000",
  font_size = 11,
  bold = FALSE,
  italic = FALSE,
  openxlsx_style = NULL,
  gt_style = NULL
)
}
\arguments{
\item{format}{formatting used for openxlsx and gt. The easiest option is using one of the predefined
formats (e.g., format_numeric()). Alternatively, pass a list with (1) a field called gt with a function for
formatting gt columns and (2) an argument passed to the numFmt field for openxlsx::createStyle. Example: list(gt = gt::fmt_auto, openxlsx = "TEXT")}

\item{background_color}{hex code for the background color}

\item{text_color}{hex code for the text color}

\item{font_size}{font size}

\item{bold}{set to TRUE for bold}

\item{italic}{set to TRUE for italic}

\item{openxlsx_style}{allows passing a custom openxlsx style. When not NULL, all other arguments will be ignored for
openxlsx exports. The object must be created with openxlsx::createStyle}

\item{gt_style}{allows passing a custom gt style. When not NULL, all other arguments will be ignored for
gt exports. Must be a function with three arguments: function(data, column, rows){ logic applying style to the specified rows in the columns }}
}
\description{
Create a new style to be applied to the body of the table.
}
\examples{
library(tablespan)
library(dplyr)
data("mtcars")

# We want to report the following table:
summarized_table <- mtcars |>
  group_by(cyl, vs) |>
  summarise(N = n(),
            mean_hp = mean(hp),
            sd_hp = sd(hp),
            mean_wt = mean(wt),
            sd_wt = sd(wt))

# Create a tablespan:
tbl <- tablespan(data = summarized_table,
                 formula = Cylinder:cyl + Engine:vs ~
                   N +
                   (`Horse Power` = Mean:mean_hp + SD:sd_hp) +
                   (`Weight` = Mean:mean_wt + SD:sd_wt),
                 title = "Motor Trend Car Road Tests",
                 subtitle = "A table created with tablespan",
                 footnote = "Data from the infamous mtcars data set.")

tbl |>
  set_style_column(columns = mean_hp,
                   style = create_style_column(bold = TRUE)) |>
  as_gt()
}
