% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deparse_formula.R
\name{deparse_formula}
\alias{deparse_formula}
\title{deparse_formula}
\usage{
deparse_formula(formula)
}
\arguments{
\item{formula}{an R formula following the pattern above}
}
\value{
nested lists for the left hand side and right hand side of the formula
}
\description{
Translates an R formula into nested lists for both the left and the right
hand side to create table rows and banners. Importantly, this function
can only parse a specific subset of formulas that allows for the following
grammar:
}
\details{
1) Braces group columns
2) Equal signs assign the left hand side as name for multiple columns
3) plus signs combine columns
}
\examples{
library(tabelle)
deparsed <- tabelle:::deparse_formula(formula =
 (`Row Name` = `Row 1` + `Row 2`) ~ `Column 1` + (`Column Banner` = `Column 2` + `Column 3`))
str(deparsed)
}
\keyword{internal}
