% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deparse_formula.R
\name{deparse_formula_partial}
\alias{deparse_formula_partial}
\title{deparse_formula_partial}
\usage{
deparse_formula_partial(
  formula_partial,
  deparsed = list(name = "_BASE_LEVEL_", entries = list())
)
}
\arguments{
\item{formula_partial}{the left or right hand side of an R formula following the pattern above}

\item{deparsed}{this function is a recursive function and makes use of the deparsed list.
Don't change this manually.}
}
\value{
a nested list with formula elements
}
\description{
Translates the left of right hand side of an R formula into a nested list. Importantly, this function
can only parse a specific subset of formulas that allows for the following
grammar:
}
\details{
1) Braces group columns
2) Equal signs assign the left hand side as name for multiple columns
3) plus signs combine columns
}
\examples{
library(tabelle)
formula <- (`Row Name` = `Row 1` + `Row 2`) ~
  `Column 1` + (`Column Banner` = `Column 2` + `Column 3`)

deparsed <- tabelle:::deparse_formula_partial(formula_partial = formula[[2]])
str(deparsed)
}
\keyword{internal}
