% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_gt.R
\name{flatten_table}
\alias{flatten_table}
\title{flatten_table}
\usage{
flatten_table(tbl)
}
\arguments{
\item{tbl}{table created with tablespan::tablespan}
}
\description{
The table header within tables created with tablespan is represented in
highly nested lists. The following function "flattens" this list to simplify
implementing the same headers in gt.
}
\examples{
library(tablespan)
library(dplyr)
data("mtcars")

summarized_table <- mtcars |>
  group_by(cyl, vs) |>
  summarise(N = n(),
            mean_hp = mean(hp),
            sd_hp = sd(hp),
            mean_wt = mean(wt),
            sd_wt = sd(wt))

tbl <- tablespan(data = summarized_table,
                 formula = (LHS = Cylinder:cyl + Engine:vs) ~
                   N +
                   (Results = (`Horse Power` = Mean:mean_hp + SD:sd_hp) +
                      (`Weight` = Mean:mean_wt + SD:sd_wt)))
str(tablespan:::flatten_table(tbl))
}
\keyword{internal}
