% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_gt.R
\name{flatten_table_partial}
\alias{flatten_table_partial}
\title{flatten_table_partial}
\usage{
flatten_table_partial(tbl_partial, flattened = list())
}
\arguments{
\item{tbl_partial}{partial of a table header}

\item{flattened}{list filled recursively}
}
\description{
Called by tablespan:::flatten_table. Recursive function that flattens the
left hand or right hand side of the table headers.
}
\examples{
library(tablespan)
library(dplyr)
data("mtcars")

summarized_table <- mtcars |>
  group_by(cyl, vs) |>
  summarise(N = n(),
            mean_hp = mean(hp),
            sd_hp = sd(hp),
            mean_wt = mean(wt),
            sd_wt = sd(wt))

tbl <- tablespan(data = summarized_table,
                 formula = (LHS = Cylinder:cyl + Engine:vs) ~
                   N +
                   (Results = (`Horse Power` = Mean:mean_hp + SD:sd_hp) +
                      (`Weight` = Mean:mean_wt + SD:sd_wt)))
str(tablespan:::flatten_table_partial(tbl$header$rhs))
}
\keyword{internal}
