% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formats.R
\name{format_date}
\alias{format_date}
\title{format_date}
\usage{
format_date(fmt = "\%Y-\%m-\%d")
}
\arguments{
\item{fmt}{date format}
}
\value{
a tablespan_format
}
\description{
Implements simple formatting for dates in tablespan.
}
