% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formats.R
\name{format_number}
\alias{format_number}
\title{format_number}
\usage{
format_number(decimals, sep_mark = ",", dec_mark = ".")
}
\arguments{
\item{decimals}{the number of decimals to show}

\item{sep_mark}{optional symbol used to separate thousands}

\item{dec_mark}{symbol used to separate decimals}
}
\value{
a tablespan_format
}
\description{
Implements simple formatting for numbers in tablespan.
}
