% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabelle.R
\name{get_col_data}
\alias{get_col_data}
\title{get_col_data}
\usage{
get_col_data(data, col_variables)
}
\arguments{
\item{data}{data set}

\item{col_variables}{variable names from formula}
}
\description{
Extracts the data that will be used as row names in the table
}
\keyword{internal}
