% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_locations.R
\name{get_locations}
\alias{get_locations}
\title{get_locations}
\usage{
get_locations(tbl, start_row, start_col)
}
\arguments{
\item{tbl}{table created with tablespan::tablespan}

\item{start_row}{row index at which the table will start}

\item{start_col}{column index at which the table will start}
}
\value{
list with locations
}
\description{
Provides row and column indices for the table elements.
}
\keyword{internal}
