% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablespan.R
\name{get_row_data}
\alias{get_row_data}
\title{get_row_data}
\usage{
get_row_data(data, row_variables)
}
\arguments{
\item{data}{data set}

\item{row_variables}{variable names from formula}
}
\description{
Extracts the data that will be used as body in the table
}
\keyword{internal}
