% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deparse_formula.R
\name{get_variables_from_list}
\alias{get_variables_from_list}
\title{get_variables_from_list}
\usage{
get_variables_from_list(deparsed_formula_element, variables = c())
}
\arguments{
\item{deparsed_formula_element}{left or right hand side of the result from
tabelle:::deparse_formula}

\item{variables}{the function is recursive and fills the variable vector}
}
\value{
a vector with the names of the variables
}
\description{
Extracts the variable names from the left or right hand side of a
deparsed formula (see ?tabelle:::deparse_formula).
}
\examples{
library(tabelle)
deparsed <- tabelle:::deparse_formula(formula =
 (`Row Name` = `Row 1` + `Row 2`) ~ `Column 1` + (`Column Banner` = `Column 2` + `Column 3`))
str(deparsed)
tabelle:::get_variables_from_list(deparsed$lhs)
}
\keyword{internal}
