% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_excel.R
\name{merge_rownames}
\alias{merge_rownames}
\title{merge_rownames}
\usage{
merge_rownames(workbook, sheet, table_data, locations, styles)
}
\arguments{
\item{workbook}{Excel workbook created with openxlsx::createWorkbook()}

\item{sheet}{name of the sheet to which the table should be written to}

\item{table_data}{data for rownames and the actual data for the body of the table}

\item{locations}{list with overview of row and col locations for different table elements}

\item{styles}{openxlsx style for the different table elements (see ?tablespan::tab_styles).
The styles element also allows applying custom styles to parts of the data shown in the
table body.}
}
\description{
Merges consecutive rownames that are identical into a common cell.
}
\keyword{internal}
