% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_tablespan.R
\name{print.Tablespan}
\alias{print.Tablespan}
\title{print.Tablespan}
\usage{
\method{print}{Tablespan}(x, digits = 2, n = 3, use_hux = require_huxtable(throw = FALSE), ...)
}
\arguments{
\item{x}{result from tablespan}

\item{digits}{number of digits to round doubles to}

\item{n}{number of rows to print}

\item{use_hux}{if set to TRUE and huxtable is installed, huxtable will be used
to print the tablespan table. This allows for styling to be printed}

\item{...}{additional arguments passed to prmatrix or huxtable (if use_hux = TRUE)}
}
\value{
nothing
}
\description{
print.Tablespan
}
\examples{
data("iris")
tbl <- tablespan(data = iris[iris$Species == "setosa", ],
          formula = Species ~ (Sepal = Sepal.Length + Sepal.Width) +
            (Petal = Petal.Length + Petal.Width))
print(tbl)
}
