% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_flextable.R
\name{require_flextable}
\alias{require_flextable}
\title{require_flextable}
\usage{
require_flextable(throw = TRUE)
}
\arguments{
\item{throw}{throw error if the package is not installed}
}
\value{
boolean or error
}
\description{
Check that flextable is installed
}
\examples{
library(tablespan)
require_flextable()
}
