% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_gt.R
\name{require_gt}
\alias{require_gt}
\title{require_gt}
\usage{
require_gt(throw = TRUE)
}
\arguments{
\item{throw}{throw error if the package is not installed}
}
\value{
boolean or error
}
\description{
Check that gt is installed
}
\examples{
library(tablespan)
require_gt()
}
