% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_hux.R
\name{require_huxtable}
\alias{require_huxtable}
\title{require_huxtable}
\usage{
require_huxtable(throw = TRUE)
}
\arguments{
\item{throw}{throw error if the package is not installed}
}
\value{
boolean or error
}
\description{
Check that huxtable is installed
}
\examples{
library(tablespan)
require_huxtable()
}
