% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_excel.R
\name{require_openxlsx}
\alias{require_openxlsx}
\title{require_openxlsx}
\usage{
require_openxlsx(throw = TRUE)
}
\arguments{
\item{throw}{throw error if the package is not installed}
}
\value{
boolean or error
}
\description{
Check that openxlsx is installed
}
\examples{
library(tablespan)
require_openxlsx()
}
