% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styles.R
\name{set_style_column}
\alias{set_style_column}
\title{set_style_column}
\usage{
set_style_column(
  tbl,
  style,
  columns = dplyr::everything(),
  rows = NULL,
  stack = TRUE
)
}
\arguments{
\item{tbl}{tablespan table}

\item{style}{the style applied to the column / cells. This style must be created with create_style_column.}

\item{columns}{the columns to style. Must be a tidyselect selector expression (e.g., starts_with("hp_"))}

\item{rows}{indices of the rows which should be styled. When set to NULL, the style is applied to all rows}

\item{stack}{When set to TRUE, the style is added on top of the existing styles. This is mostly relevant
for openxlsx. When set to FALSE, the new style replaces all previous styling.}
}
\value{
the tablespan table with added styles
}
\description{
Change the style of a column or single cells within columns.
}
\details{
To change the style of the data shown in the body of the table,
}
\examples{
library(tablespan)
library(dplyr)
data("mtcars")

# We want to report the following table:
summarized_table <- mtcars |>
  group_by(cyl, vs) |>
  summarise(N = n(),
            mean_hp = mean(hp),
            sd_hp = sd(hp),
            mean_wt = mean(wt),
            sd_wt = sd(wt))

# Create a tablespan:
tbl <- tablespan(data = summarized_table,
                 formula = Cylinder:cyl + Engine:vs ~
                   N +
                   (`Horse Power` = Mean:mean_hp + SD:sd_hp) +
                   (`Weight` = Mean:mean_wt + SD:sd_wt),
                 title = "Motor Trend Car Road Tests",
                 subtitle = "A table created with tablespan",
                 footnote = "Data from the infamous mtcars data set.")

tbl |>
  set_style_column(columns = mean_hp,
                   style = create_style_column(bold = TRUE)) |>
  as_gt()
}
