% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styles.R
\name{set_style_header_cells}
\alias{set_style_header_cells}
\title{set_style_header_cells}
\usage{
set_style_header_cells(tbl, openxlsx_style)
}
\arguments{
\item{tbl}{tablespan table}

\item{openxlsx_style}{style used when exporting to openxlsx}
}
\value{
the tablespan table with added styles
}
\description{
Set the style used for the cells in the openxlsx export. This function
is used to create the borders around cells in openxlsx.
}
\details{
- openxlsx_style must be a style object created with openxlsx::createStyle. This style
will then be applied to the header

All functions that start with "set_style_" completely replace existing styling.
}
\examples{
library(tablespan)
library(dplyr)
data("mtcars")

# We want to report the following table:
summarized_table <- mtcars |>
  group_by(cyl, vs) |>
  summarise(N = n(),
            mean_hp = mean(hp),
            sd_hp = sd(hp),
            mean_wt = mean(wt),
            sd_wt = sd(wt))

# Create a tablespan:
tbl <- tablespan(data = summarized_table,
                 formula = Cylinder:cyl + Engine:vs ~
                   N +
                   (`Horse Power` = Mean:mean_hp + SD:sd_hp) +
                   (`Weight` = Mean:mean_wt + SD:sd_wt),
                 title = "Motor Trend Car Road Tests",
                 subtitle = "A table created with tablespan",
                 footnote = "Data from the infamous mtcars data set.")

wb <- tbl |>
  set_style_header_cells(
    openxlsx_style = openxlsx::createStyle(
       fontSize = 11,
       halign = "center",
       border = "Bottom",
       borderColour = "#000000",
       borderStyle = "thin",
       textDecoration = "bold"
     )) |>
  as_excel()
# save workbook to see the effect
}
