% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_gt.R
\name{to_gt}
\alias{to_gt}
\title{to_gt}
\usage{
to_gt(tbl, groupname_col = FALSE, ...)
}
\arguments{
\item{tbl}{table created with tablespan::tablespan}

\item{groupname_col}{Provide column names to group data. See ?gt::gt for more
details.}

\item{...}{additional arguments passed to gt::gt().}
}
\description{
Translates a table created with tablespan to a great table (gt). See https://gt.rstudio.com/.
}
\details{
Tablespan itself does not provide any printing of tables as HTML table. However,
with to_gt, tablespan can be translated to a great table which provides html and
LaTeX output.
}
\examples{
library(tablespan)
library(dplyr)
data("mtcars")

summarized_table <- mtcars |>
  group_by(cyl, vs) |>
  summarise(N = n(),
            mean_hp = mean(hp),
            sd_hp = sd(hp),
            mean_wt = mean(wt),
            sd_wt = sd(wt))

tbl <- tablespan(data = summarized_table,
                 formula = (LHS = Cylinder:cyl + Engine:vs) ~
                   N +
                   (Results = (`Horse Power` = Mean:mean_hp + SD:sd_hp) +
                      (`Weight` = Mean:mean_wt + SD:sd_wt)))

gt_tbl <- to_gt(tbl)
print(gt_tbl)
}
