% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_table.R
\name{write_bt}
\alias{write_bt}
\title{write_bt}
\usage{
write_bt(
  tbl,
  workbook = openxlsx::createWorkbook(),
  sheet = "BasicTable",
  start_row = 1,
  start_col = 1,
  header_style = openxlsx::createStyle(fontSize = 14, halign = "center", border =
    "Bottom", borderColour = openxlsx_getOp("borderColour", "black"), borderStyle =
    openxlsx_getOp("borderStyle", "double"), textDecoration = "bold")
)
}
\arguments{
\item{tbl}{table created with basicTables::bt}

\item{workbook}{Excel workbook created with openxlsx::createWorkbook()}

\item{sheet}{name of the sheet to which the table should be written to}

\item{start_row}{row at which to start the table}

\item{start_col}{column at which to start the table}

\item{header_style}{openxlsx style for the header}
}
\description{
Write a basic table to an excel workbook.
}
\examples{
library(basciTables)
data("iris")

tbl <- bt(data = iris[iris$Species == "setosa", ],
          formula = Species ~ (Sepal = Sepal.Length + Sepal.Width) +
            (Petal = (Width = Petal.Length) + Petal.Width))

wb <- write_bt(tbl = tbl)

# saveWorkbook(wb, "iris.xlsx")
}
