% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_excel.R
\name{write_footnote}
\alias{write_footnote}
\title{write_footnote}
\usage{
write_footnote(tbl, workbook, sheet, locations, styles)
}
\arguments{
\item{tbl}{table created with tablespan::tablespan}

\item{workbook}{Excel workbook created with openxlsx::createWorkbook()}

\item{sheet}{name of the sheet to which the table should be written to}

\item{locations}{list with overview of row and col locations for different table elements}

\item{styles}{openxlsx style for the different table elements (see ?tablespan::tab_styles).
The styles element also allows applying custom styles to parts of the data shown in the
table body.}
}
\description{
Writes the footnote to the workbook.
}
\keyword{internal}
