% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_excel.R
\name{write_header}
\alias{write_header}
\title{write_header}
\usage{
write_header(workbook, sheet, header, table_data, locations, styles)
}
\arguments{
\item{workbook}{Excel workbook created with openxlsx::createWorkbook()}

\item{sheet}{name of the sheet to which the table should be written to}

\item{header}{header specification from tablespan table}

\item{table_data}{data for rownames and the actual data for the body of the table}

\item{locations}{list with overview of row and col locations for different table elements}

\item{styles}{openxlsx style for the different table elements (see ?tablespan::tab_styles).
The styles element also allows applying custom styles to parts of the data shown in the
table body.}
}
\description{
Writes the header (column names and names for the rownames) to the workbook.
}
\keyword{internal}
