% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_excel.R
\name{write_header_entry}
\alias{write_header_entry}
\title{write_header_entry}
\usage{
write_header_entry(
  workbook,
  sheet,
  header_entry,
  max_level,
  start_row,
  start_col,
  header_style,
  vline_style
)
}
\arguments{
\item{workbook}{Excel workbook created with openxlsx::createWorkbook()}

\item{sheet}{name of the sheet to which the table should be written to}

\item{header_entry}{specific header enty to write to the workbook}

\item{max_level}{the highest level of header entries}

\item{start_row}{integer specifying row to write to}

\item{start_col}{integer specifying column to write to}

\item{header_style}{openxlsx style for the header}

\item{vline_style}{openxlsx style for the vertical lines in the header}
}
\description{
Recursive function writing the header elements (column names and names for the rownames)
to the workbook.
}
\keyword{internal}
