% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_flextable.R
\name{as_flextable.Tablespan}
\alias{as_flextable.Tablespan}
\title{as_flextable}
\usage{
\method{as_flextable}{Tablespan}(x, theme = flextable::theme_booktabs, ...)
}
\arguments{
\item{x}{table created with tablespan::tablespan}

\item{theme}{a theme to apply to the flextable. Use one of the flextable::theme_* functions}

\item{...}{additional arguments passed to flextable::as_flextable}
}
\value{
flextable that can be further adapted with the gt package.
}
\description{
Translates a table created with tablespan to a flextable. See <https://ardata-fr.github.io/flextable-book/>.
}
\details{
Flextable is an extremely versatile table creator for R with great support to export to a variety of formats.
}
\examples{
library(tablespan)
library(dplyr)
data("mtcars")

summarized_table <- mtcars |>
  group_by(cyl, vs) |>
  summarise(N = n(),
            mean_hp = mean(hp),
            sd_hp = sd(hp),
            mean_wt = mean(wt),
            sd_wt = sd(wt))

tbl <- tablespan(data = summarized_table,
                 formula = (LHS = Cylinder:cyl + Engine:vs) ~
                   N +
                   (Results = (`Horse Power` = Mean:mean_hp + SD:sd_hp) +
                      (`Weight` = Mean:mean_wt + SD:sd_wt)))
if(require_flextable(throw = FALSE)){
  library(flextable)
  flex_tbl <- as_flextable(tbl)
  flex_tbl
}
}
