% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_googlesheet_request.R
\name{fake_gs4_dribble}
\alias{fake_gs4_dribble}
\title{fake_gs4_dribble}
\usage{
fake_gs4_dribble()
}
\value{
fake dribble
}
\description{
Creates a fake googlesheets4 dribble to use as a placeholder in the as_googlesheets_request function.
}
\examples{
library(tablespan)
fake_gs4_dribble()
}
