% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formats.R
\name{format_number}
\alias{format_number}
\title{format_number}
\usage{
format_number(decimals = 2, sep_mark = ",", dec_mark = ".")
}
\arguments{
\item{decimals}{the number of decimals to show}

\item{sep_mark}{optional symbol used to separate thousands}

\item{dec_mark}{symbol used to separate decimals}
}
\value{
a list with styles for gt and openxlsx
}
\description{
Implements simple formatting for numbers in gt and excel exports of tablespan.
}
\examples{
library(tablespan)
library(dplyr)
data("mtcars")

# We want to report the following table:
summarized_table <- mtcars |>
  group_by(cyl, vs) |>
  summarise(N = n(),
            mean_hp = mean(hp),
            sd_hp = sd(hp),
            mean_wt = mean(wt),
            sd_wt = sd(wt))

# Create a tablespan:
tbl <- tablespan(data = summarized_table,
                 formula = Cylinder:cyl + Engine:vs ~
                   N +
                   (`Horse Power` = Mean:mean_hp + SD:sd_hp) +
                   (`Weight` = Mean:mean_wt + SD:sd_wt),
                 title = "Motor Trend Car Road Tests",
                 subtitle = "A table created with tablespan",
                 footnote = "Data from the infamous mtcars data set.")

tbl |>
  style_column(columns = mean_hp,
              format = format_number(decimals = 5)) |>
  as_gt()
}
