% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deparse_formula.R
\name{get_variables}
\alias{get_variables}
\title{get_variables}
\usage{
get_variables(deparsed_formula)
}
\arguments{
\item{deparsed_formula}{result from tabelle:::deparse_formula}
}
\value{
a list with the names of the variables that build the rows and columns
}
\description{
Extracts the variable names from a deparsed formula (see ?tabelle:::deparse_formula).
}
\examples{
library(tabelle)
deparsed <- tabelle:::deparse_formula(formula =
 (`Row Name` = `Row 1` + `Row 2`) ~ `Column 1` + (`Column Banner` = `Column 2` + `Column 3`))
str(deparsed)
tabelle:::get_variables(deparsed)
}
\keyword{internal}
