% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_basic_table.R
\name{print.Tablespan}
\alias{print.Tablespan}
\title{print.Tablespan}
\usage{
\method{print}{Tablespan}(x, digits = 2, n = 3, ...)
}
\arguments{
\item{x}{result from tablespan}

\item{digits}{number of digits to round doubles to}

\item{n}{number of rows to print}

\item{...}{additional arguments passed to prmatrix}
}
\value{
nothing
}
\description{
print.Tablespan
}
\examples{
data("iris")
tbl <- tablespan(data = iris[iris$Species == "setosa", ],
          formula = Species ~ (Sepal = Sepal.Length + Sepal.Width) +
            (Petal = Petal.Length + Petal.Width))
print(tbl)
}
