% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_googlesheet_request.R
\name{require_googlesheets4}
\alias{require_googlesheets4}
\title{Check if googlesheets4 package is available}
\usage{
require_googlesheets4(throw = TRUE)
}
\arguments{
\item{throw}{logical. If TRUE (default), the function will throw an error if
googlesheets4 is not available. If FALSE, it will return FALSE instead.}
}
\value{
logical. Returns TRUE if googlesheets4 is available, FALSE if it's not
        available and throw=FALSE.
}
\description{
This function checks if the googlesheets4 package is available.
}
\examples{
\dontrun{
# Check if googlesheets4 is available, throw error if not
require_googlesheets4()

# Check if googlesheets4 is available, return FALSE if not
require_googlesheets4(throw = FALSE)
}
}
