% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_table.R
\name{write_bt}
\alias{write_bt}
\title{write_bt}
\usage{
write_bt(
  tbl,
  workbook = openxlsx::createWorkbook(),
  sheet = "BasicTable",
  start_row = 1,
  start_col = 1,
  styles = bt_styles()
)
}
\arguments{
\item{tbl}{table created with basicTables::bt}

\item{workbook}{Excel workbook created with openxlsx::createWorkbook()}

\item{sheet}{name of the sheet to which the table should be written to}

\item{start_row}{row at which to start the table}

\item{start_col}{column at which to start the table}

\item{styles}{openxlsx style for the different table elements (see ?basicTables::bt_styles)}
}
\description{
Write a basic table to an excel workbook.
}
\examples{
library(basicTables)
data("iris")

tbl <- bt(data = iris[iris$Species == "setosa", ],
          formula = Species ~ (Sepal = Sepal.Length + Sepal.Width) +
            (Petal = (Width = Petal.Length) + Petal.Width))

wb <- write_bt(tbl = tbl)

# saveWorkbook(wb, "iris.xlsx")
}
